codreviewer/requirements.txt
Flask==3.1.3
gunicorn==25.1.0
pycryptodome==3.23.0

